REMOVE: Delete test_embedded_files.py as its tests have been integrated into test_files.py, streamlining the test suite and improving organization.
