+++ tests/generic/test_files.py
@@ -7,9 +7,18 @@ from pathlib import Path
 
 import pytest
 
-from pypdf import PdfReader
+from pypdf import PdfReader, PdfWriter
 from pypdf.errors import PdfReadError
-from pypdf.generic import ByteStringObject, DictionaryObject, EmbeddedFile, NameObject
+from pypdf.generic import (
+    ByteStringObject,
+    DictionaryObject,
+    EmbeddedFile,
+    NameObject,
+    NullObject,
+    NumberObject,
+    TextStringObject,
+    create_string_object,
+)
 from tests import get_data_from_url
 
 TESTS_ROOT = Path(__file__).parent.parent.resolve()
@@ -109,3 +118,214 @@ def test_embedded_file__kids():
     )
     attachments = list(EmbeddedFile._load(reader.root_object))
     assert attachments == []
+
+
+def test_embedded_file_name_is_read_only():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    assert embedded_file.name == "test.txt"
+
+    with pytest.raises(AttributeError):
+        embedded_file.name = "new_name.txt"
+
+
+def test_embedded_file_alternative_name_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    embedded_file.alternative_name = TextStringObject("Alternative Name")
+    assert embedded_file.alternative_name == "Alternative Name"
+
+    embedded_file.alternative_name = None
+    if NameObject("/UF") in embedded_file.pdf_object:
+        assert embedded_file.pdf_object[NameObject("/UF")] == NullObject()
+    if NameObject("/F") in embedded_file.pdf_object:
+        assert embedded_file.pdf_object[NameObject("/F")] == NullObject()
+    assert embedded_file.alternative_name is None
+
+    pdf_string = TextStringObject("PDF String")
+    embedded_file.alternative_name = pdf_string
+    assert embedded_file.alternative_name == "PDF String"
+
+
+def test_embedded_file_alternative_name_both_f_and_uf():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    embedded_file.pdf_object[NameObject("/F")] = create_string_object("original_f")
+    embedded_file.pdf_object[NameObject("/UF")] = create_string_object("original_uf")
+
+    embedded_file.alternative_name = TextStringObject("new_name")
+    assert embedded_file.pdf_object[NameObject("/F")] == create_string_object("new_name")
+    assert embedded_file.pdf_object[NameObject("/UF")] == create_string_object("new_name")
+    assert embedded_file.alternative_name == "new_name"
+
+    embedded_file.alternative_name = None
+    assert embedded_file.pdf_object[NameObject("/F")] == NullObject()
+    assert embedded_file.pdf_object[NameObject("/UF")] == NullObject()
+    assert embedded_file.alternative_name is None
+
+
+def test_embedded_file_description_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    embedded_file.description = TextStringObject("Test Description")
+    assert embedded_file.description == "Test Description"
+
+    embedded_file.description = None
+    assert embedded_file.pdf_object[NameObject("/Desc")] == NullObject()
+
+    pdf_string = TextStringObject("PDF Description")
+    embedded_file.description = pdf_string
+    assert embedded_file.description == "PDF Description"
+
+
+def test_embedded_file_subtype_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    embedded_file.subtype = NameObject("/text/plain")
+    assert embedded_file.subtype == "/text/plain"
+
+    embedded_file.subtype = None
+    assert embedded_file._embedded_file[NameObject("/Subtype")] == NullObject()
+
+    name_obj = NameObject("/application#2Fjson")
+    embedded_file.subtype = name_obj
+    assert embedded_file.subtype == "/application#2Fjson"
+
+
+def test_embedded_file_size_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    embedded_file.size = NumberObject(1024)
+    assert embedded_file.size == 1024
+
+    embedded_file.size = None
+    assert embedded_file._ensure_params[NameObject("/Size")] == NullObject()
+
+    num_obj = NumberObject(2048)
+    embedded_file.size = num_obj
+    assert embedded_file.size == 2048
+
+
+def test_embedded_file_size_getter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    # Test null case
+    embedded_file._ensure_params[NameObject("/Size")] = NullObject()
+    assert embedded_file.size is None
+
+    embedded_file._ensure_params[NameObject("/Size")] = NumberObject(4096)
+    retrieved_size = embedded_file.size
+    assert retrieved_size == 4096
+
+
+def test_embedded_file_creation_date_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    # Test with datetime object
+    test_date = datetime.datetime(2023, 1, 1, 12, 0, 0)
+    embedded_file.creation_date = test_date
+    assert embedded_file.creation_date == test_date
+
+    embedded_file.creation_date = None
+    assert embedded_file._ensure_params[NameObject("/CreationDate")] == NullObject()
+
+    # Test with TextStringObject
+    date_string = TextStringObject("D:20230101120000")
+    embedded_file.creation_date = date_string
+    assert embedded_file.creation_date is not None
+
+
+def test_embedded_file_modification_date_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    # Test with datetime object
+    test_date = datetime.datetime(2023, 1, 2, 12, 0, 0)
+    embedded_file.modification_date = test_date
+    assert embedded_file.modification_date == test_date
+
+    embedded_file.modification_date = None
+    assert embedded_file._ensure_params[NameObject("/ModDate")] == NullObject()
+
+    # Test with TextStringObject
+    date_string = TextStringObject("D:20230102120000")
+    embedded_file.modification_date = date_string
+    assert embedded_file.modification_date is not None
+
+
+def test_embedded_file_checksum_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    checksum_bytes = ByteStringObject(b"checksum_value")
+    embedded_file.checksum = checksum_bytes
+    assert embedded_file.checksum == b"checksum_value"
+
+    embedded_file.checksum = None
+    assert embedded_file._ensure_params[NameObject("/CheckSum")] == NullObject()
+
+    byte_string = ByteStringObject(b"pdf_checksum")
+    embedded_file.checksum = byte_string
+    assert embedded_file.checksum == b"pdf_checksum"
+
+
+def test_embedded_file_associated_file_relationship_setter():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    embedded_file.associated_file_relationship = NameObject("/Data")
+    assert embedded_file.associated_file_relationship == "/Data"
+
+
+def test_embedded_file_setters_integration():
+    writer = PdfWriter()
+    writer.add_blank_page(100, 100)
+
+    embedded_file = writer.add_attachment("test.txt", b"Hello, World!")
+    embedded_file.alternative_name = TextStringObject("Alternative Name")
+    embedded_file.description = TextStringObject("Test Description")
+    embedded_file.subtype = NameObject("/text/plain")
+    embedded_file.size = NumberObject(13)
+    creation_date = datetime.datetime(2023, 1, 1, 12, 0, 0)
+    embedded_file.creation_date = creation_date
+    modification_date = datetime.datetime(2023, 1, 2, 12, 0, 0)
+    embedded_file.modification_date = modification_date
+    embedded_file.checksum = ByteStringObject(b"checksum123")
+    embedded_file.associated_file_relationship = NameObject("/Data")
+
+    pdf_bytes = BytesIO()
+    writer.write(pdf_bytes)
+
+    reader = PdfReader(pdf_bytes)
+    assert "test.txt" in reader.attachments
+
+
+def test_embedded_file_null_object_handling():
+    writer = PdfWriter()
+    embedded_file = writer.add_attachment("test.txt", b"content")
+
+    embedded_file.alternative_name = TextStringObject("Name")
+    embedded_file.description = TextStringObject("Description")
+    embedded_file.subtype = NameObject("/text/plain")
+    embedded_file.size = NumberObject(1024)
+    embedded_file.checksum = ByteStringObject(b"checksum")
+
+    embedded_file.alternative_name = None
+    embedded_file.description = None
+    embedded_file.subtype = None
+    embedded_file.size = None
+    embedded_file.checksum = None
+
+    assert embedded_file.alternative_name is None
+    assert embedded_file.description is None
+    assert embedded_file.subtype is None
+    assert embedded_file.size is None
+    assert embedded_file.checksum is None
